Move and rename specs
